requirements.txt
streamlit~=1.50.0
langchain-openai~=0.3.34
langchain~=0.3.27
RapidFuzz~=3.14.1
langgraph~=0.6.8
python-dotenv~=1.1.1
mysql-connector-python~=9.4.0